# Optimize memory usage with DBString and DBArray