--- a/Makefile.am
+++ b/Makefile.am
@@ -43,6 +43,7 @@ liblcf_la_LDFLAGS = \
 	-no-undefined
 liblcf_la_SOURCES = \
 	src/data.cpp \
+	src/dbstring.cpp \
 	src/encoder.cpp \
 	src/ini.cpp \
 	src/inireader.cpp \
@@ -201,6 +202,7 @@ liblcf_la_SOURCES = \
 
 lcfinclude_HEADERS = \
 	src/lcf/data.h \
+	src/lcf/dbstring.h \
 	src/lcf/encoder.h \
 	src/lcf/enum_tags.h \
 	src/lcf/flag_set.h \