--- a/Makefile.am
+++ b/Makefile.am
@@ -43,7 +43,7 @@ liblcf_la_LDFLAGS = \
 	-no-undefined
 liblcf_la_SOURCES = \
 	src/data.cpp \
-	src/dbstring.cpp \
+	src/dbarray.cpp \
 	src/encoder.cpp \
 	src/ini.cpp \
 	src/inireader.cpp \
@@ -202,6 +202,8 @@ liblcf_la_SOURCES = \
 
 lcfinclude_HEADERS = \
 	src/lcf/data.h \
+	src/lcf/dbarray.h \
+	src/lcf/dbarrayalloc.h \
 	src/lcf/dbstring.h \
 	src/lcf/encoder.h \
 	src/lcf/enum_tags.h \