--- a/src/dbstring.cpp
+++ b/src/dbstring.cpp
@@ -1,7 +1,5 @@
 #include "lcf/dbstring.h"
-#include <unordered_map>
 #include <memory>
-#include <iostream>
 
 namespace lcf {