--- a/src/generated/lcf/rpg/chipset.h
+++ b/src/generated/lcf/rpg/chipset.h
@@ -16,6 +16,7 @@
 #include <stdint.h>
 #include <string>
 #include <vector>
+#include "lcf/dbstring.h"
 #include "lcf/enum_tags.h"
 #include <ostream>
 #include <type_traits>
@@ -39,7 +40,7 @@ namespace rpg {
 		Chipset();
 		void Init();
 		int ID = 0;
-		std::string name;
+		DBString name;
 		std::string chipset_name;
 		std::vector<int16_t> terrain_data;
 		std::vector<uint8_t> passable_data_lower;