--- a/src/generated/lcf/rpg/chipset.h
+++ b/src/generated/lcf/rpg/chipset.h
@@ -14,7 +14,6 @@
 
 // Headers
 #include <stdint.h>
-#include <string>
 #include <vector>
 #include "lcf/dbstring.h"
 #include "lcf/enum_tags.h"
@@ -41,7 +40,7 @@ namespace rpg {
 		void Init();
 		int ID = 0;
 		DBString name;
-		std::string chipset_name;
+		DBString chipset_name;
 		std::vector<int16_t> terrain_data;
 		std::vector<uint8_t> passable_data_lower;
 		std::vector<uint8_t> passable_data_upper;