--- a/src/generated/lcf/rpg/eventcommand.h
+++ b/src/generated/lcf/rpg/eventcommand.h
@@ -14,8 +14,8 @@
 
 // Headers
 #include <stdint.h>
-#include <string>
 #include <vector>
+#include "lcf/dbstring.h"
 #include "lcf/enum_tags.h"
 #include <ostream>
 #include <type_traits>
@@ -179,7 +179,7 @@ namespace rpg {
 
 		int32_t code = 0;
 		int32_t indent = 0;
-		std::string string;
+		DBString string;
 		std::vector<int32_t> parameters;
 	};
 	inline std::ostream& operator<<(std::ostream& os, EventCommand::Code code) {