--- a/src/generated/lcf/rpg/eventcommand.h
+++ b/src/generated/lcf/rpg/eventcommand.h
@@ -13,8 +13,8 @@
 #define LCF_RPG_EVENTCOMMAND_H
 
 // Headers
+#include <lcf/dbarray.h>
 #include <stdint.h>
-#include <vector>
 #include "lcf/dbstring.h"
 #include "lcf/enum_tags.h"
 #include <ostream>
@@ -180,7 +180,7 @@ namespace rpg {
 		int32_t code = 0;
 		int32_t indent = 0;
 		DBString string;
-		std::vector<int32_t> parameters;
+		DBArray<int32_t> parameters;
 	};
 	inline std::ostream& operator<<(std::ostream& os, EventCommand::Code code) {
 		os << static_cast<std::underlying_type_t<decltype(code)>>(code);