Depends on: https://github.com/EasyRPG/Player/pull/2280


DBString
------------

* Intended for storing read only database data.
* Optimized for space, `sizeof(DBString) == sizeof(char*)`
* Optimized for the case of many empty strings, no small string optimization as used in C++11 std::string.
* Stores the size in the buffer, pascal style
* Custom memory allocator for compact memory usage, fast loading, and optimal cache locality (coming soon)

DBarray
--------

* Same storage policy as DBString
* Replaces std::vector

Performance Measurements
------------------------------

`bench_readldb` max memory usage

| Game | Ldb Size | Master | PR w/o allocator | PR w/ allocator |
| -- | -- | -- | -- | -- |
| HH3 | 16MB | 135.7MB | 94.7MB | TBD |
| Heroes Realm | 27MB | 301.6MB | 204.3MB | TBD | 
| Violated Heroine | 11MB | 75.9MB | 54.6MB | TBD |
| Yume 2kki | 1.4MB | 8.0MB | 6.0MB | TBD

`bench_readldb` average runtime

| Game | Master | PR w/o allocator | PR w/ allocator |
| -- | -- | -- | -- |
| HH3 | 0.522s | 0.522s | TBD |
| Heroes Realm | 1.09s | 1.09s | TBD | 
| Violated Heroine | 0.323s | 0.323s | TBD |
| Yume 2kki | 0.045s | 0.046s | TBD |

Notes about Editor
------------------------

`DBString` is optimal for read-only database usage of Player but not necessarily for Editor. For editor, maybe we want to still use `std::string` for all strings. Perhaps an `EDITOR` build mode of liblcf?

Given that Editor is not being worked on, I'm not super interested in putting too much effort on this side.

Todo
------

- [x] Write unit tests for string types
- [x] Retest liblcf bit-perfect read-writing
- [x] Final review what should use DBString and what should not.
- [x] Test Player
- [x] ~Write a custom memory allocator for DBString and benchmark it~
- [x] Should we use a third party string_view like gsl or nonstd?
- [ ] Best way to handle libfmt support / hacks?
- [x] Make `DBString` move only, or at least make copies reference same underlying string using allocator?
- [x] ~`DBString` from database should be read only and exist for the lifetime of the game, maps however come and go. How to reconcile with allocator? Maybe DB and Map arenas?~
- [x] ~Add a "use std::string" for everything build mode for editor and other interactive tools.~