@fmatthew5876 2020-08-16
> Will it still be possible to allocate a new DBString and overwrite an existing DBString?
> Otherwise the editor will have a huge problem because it can't write data anymore ^^'.
> (lets ignore any speed impacts here, the editor will not run on slow hardware)

Right now DBString just does `malloc()` so yes, but once I put in the custom allocator that won't be the case.

That's why I'm suggesting either a separate build mode, or some template way of creating a `std::string` only database.

Essentially we need "fast read-only player mode" and "slow read/write editor mode" for liblcf somehow.