@fmatthew5876 2020-08-17
I pushed an experiment with deduplication. Basically each string is refcounted and we look it up in a hash table.

Unfortunately the results are not great. We get a meager savings in memory and a slower load time. For the memory, I can assume that even though a lot of strings are repeated, the memory overhead of the hash table dwarfs the savings. For the runtime, I could probably improve things by using a faster hash table.

| Game | Max Mem Usage | Runtime
| -- | -- |
| Yume 2kki | 6Mb | 0.73s |
| HH3 | 90.7MB | 0.578s |
| HeroesRealm | 1992.MB | 1.18s |

Overall looks like deduplication of strings is not worth it :-1: