@fmatthew5876 2020-08-17
And now lets try taking out the hash table deduplication, but keeping the reference counting mechanism for copies.


| Game | Max Mem Usage | 
| -- | -- |
| Yume 2kki | 6.1Mb |
| HH3 | 95.2MB |
| HeroesRealm | 203.8MB |

Memory usage is slightly up across the board. This is a also a loser. Not surprising, since we have a lot of strings and an extra 4 bytes for each one still adds up.