@fmatthew5876 2020-08-17
One more important test. What are the memory usage differences if we take events out of the picture?

Max memory usage Toop::event_commands removed
| Game | Master | PR |
| -- | -- | -- |
| HH3 | 10.9MB | 8.7MB |
| Heroes Realm | 9.3MB | 8.1MB |
| Violated Heroine | 73.9MB | 52.8MB |
| Yume 2kki | 8.0MB | 6.0MB |

Max memory usage Toop::event_commands and CommonEvent::event_commands removed

| Game | Master | PR |
| -- | -- | -- |
| HH3 | 4.9MB | 4.5MB | 8.7MB |
| Heroes Realm | 6.3MB | 6.0MB |
| Violated Heroine | 5.6MB | 5.0MB |
| Yume 2kki | 2.4MB | 2.0MB |

So the conclusion here is that for real wins, we have to either compress or not keep all these event data into memory at once. As has been suggested in the past, I agree probably some caching mechanism would make sense.

For battles this is easy, we have several frames of battle transition where we could load the battle info from disk and keep it in an LRU cache.

For common events not so much, since they can trigger at any time and even 1 frame delay could break a game.

As was also suggested before, compressing the data would help a lot. Basically keeping the event code in it's binary form and decoding it on the fly as the events execute.