@fmatthew5876 2020-08-18
> How are the savings when you keep the RPG::eventcommand as is and only make the parameter vector uint8_t with the raw BER data? Should also safe some memory this way and keep the API mostly the same.

I did that here https://github.com/EasyRPG/liblcf/issues/255#

But check out the new post if you replace the entire event stream by `uint8_t` ...


> Another 4 bytes: Will it break games when the code and indent are int16_t? The values would all fit.

I think this would be fine, but I would use `uint16_t`.
Not sure what the indent limit is in RPG_RT, but 255 indents seems like way too many.

A better option could be 24 bytes for the code and 8 bytes for the indent. 255 is the sentinal for `subcommand_path`.. That leaves the code space wide open millions of commands. A complex game building lots of custom functions could get stuck on 64K and we might want to use namespace like gaps in numbers for different patchsets, giving them room to expand without collision.

Another way to reduce 16 bytes, add a `DBArray` class which has the size embedded in the buffer like string. On 64 bit platforms we would reduce from a 24 byte std::vector to an 8 byte `DBArray`. This data structure would also help reduce memory size of items and animations, which also contain lots of sub-arrays.

Even better, make some `EventRunner` class. Store the entire array of all event command parameters in a single vector. Then use a `int32_t` offset in `EventCommand` to reference into it. Not only saving 4 bytes, but also saving all the memory allocation overhead of all the individual parameter arrays.