@fmatthew5876 2020-08-18
I've switched StringView to use this implementation:

https://github.com/martinmoene/string-view-lite

We now distribute `include/lcf/third_party` for third party header files we use in liblcf. The intention is that all third party types we put in here get wrapped in `lcf::` namespace.

Also wrote unit tests.

I'm really not thrilled about this:
https://github.com/fmatthew5876/Player/blob/string/src/string_view.h#L33

Basically, enable fmtlib printing support without including more than `<fmt/core.h>` you need to write a `to_string_view()` function which returns the `fmt::string_view` type. Since liblcf doesn't depend on fmt, I hacked it into the player header.

It's ugly, but it works and probably necessary if we don't want to add the dependency at the lcf level.

Player PR is up to date and building against this one.

This is ready for review.