@fmatthew5876 2020-08-19
Added a few more small improvements.

This PR is done.

The next step is `DBArray` which provides even more memory savings. That one requires deep refactoring of liblcf parsing to support a new array type, so it has to be in the next PR.